visible, uninterruptable teleports

+++ src/commonMain/kotlin/at/gnu/gridz/GridzTile.kt
@@ -7,10 +7,10 @@ sealed class GridzTile(open val x: Int, open val y: Int) {
 }
 
 class Empty(override val x: Int, override val y: Int) : GridzTile(x, y) {
-    var lit: Int = 0
+    var lit: Long = 0L
 
     companion object {
-        const val LIT_TIME = 50000
+        const val LIT_TIME = 50000L
     }
 }
